requirements.txt
pyyaml==5.3
redis==3.4.1
xpinyin==0.5.6
requests==2.22.0
useragentxxc==1.1.0